fix(desktop): Point::new takes i32, not f64
uiautomation 0.25's Point::new signature is (i32, i32); the previous
x as f64 / y as f64 casts came from an older docs.rs read.

+++ subpackage/cli/packages/desktop/src-tauri/src/computer_use.rs
@@ -313,7 +313,7 @@ mod ax_windows {
     pub fn node_at_point(x: i32, y: i32, max_depth: u32) -> Result<AxNode, String> {
         let automation = UIAutomation::new().map_err(|e| e.to_string())?;
         let element = automation
-            .element_from_point(Point::new(x as f64, y as f64))
+            .element_from_point(Point::new(x, y))
             .map_err(|e| e.to_string())?;
         Ok(convert(&automation, &element, max_depth))
     }
